refacotr the header action buttons
